Setup integration test for `blitz new`
Mocking out fs is not compatible with oclif (see https://github.com/oclif/oclif/issues/182).
That's why I've decided to use a temporary directory instead. AS these are deleted on startup (see https://www.2daygeek.com/automatically-delete-clean-up-tmp-directory-folder-contents-in-linux/), this should not be a big problem.
What is a big problem though, is the duration of these tests: Since it *really runs blitz new*, it takes a long time to run (which is obvious, since it needs to download the entirety of the internet for dependencies ;D).
There should be a way to differentiate between integration/e2e tests and unit tests so that the long-running ones can be omitted during dev. But that's out of scope for this PR.
The next.js repository can serve as a good inspiration on how to achieve that separation.