Refactor Template dependency version replacement
The code for fetching the latest dependency versions seemed a bit convoluted to me, so I tried to clean it up.
I hope this is easier to read for the rest of you, as well.

There's one performance downside introduced: Fetching the latest dependencies doesn't occur in parallel, but sequentally now. This shouldnt have a noticeable impact though, considering the nature of JS' event loop.
It makes the code a lot easier to read though :)