Prevent test breaking on version change
Running the test in-between version change and publish will break, since the version from package.json is new, but it's not in the registry.
So we shouldn't really test against the version in package.json, as that's not the one we intend to install anyways! We really want to install the latest one.

So I've changed the test to reflect this - there's just one simple problem:
By fetching the latest version from NPM, we basically repeat the implementation. Isn't this repetitive, shouldn't we just import the implemented function into our tests? I don't think that's a good Idea, since that would couple our tests with implementation and bugs in that part of the code wouldnt be caught.
So that's why I chose duplication here.