Add instanceof type guards for error checking
With instanceof type guards TypeScript is able to
infer the type of the error. This is only possible
because the custom errors extend the default 'Error'
type. Therefore it is not needed to cast an error
to 'any'. As an extra addition I've added a union type
for the different custom errors that are build in.